// app/networking/interface.ts
export type Episode = {
  id: number
  name: string
  air_date: string
  episode: string
  characters: string[]
  url: string
  created: string
}

export type AllEpisodes = {
  info: {
    count: number
    pages: number
    next: string
    prev: string | null
  }
  results: [
    {
      id: number
      name: string
      air_date: string
      episode: string
      characters: string[]
      url: string
      created: string
    }
  ]
}

export type Character = {
  id: number
  name: string
  status: string
  species: string
  type: string
  gender: string
  origin: {
    name: string
    url: string
  }
  location: {
    name: string
    url: string
  }
  image: string
  episode: string[]
  url: string
  created: string
}

export type MultipleCharacters = [
  {
    id: number
    name: string
    status: string
    species: string
    type: string
    gender: string
    origin: {
      name: string
      url: string
    }
    location: {
      name: string
      url: string
    }
    image: string
    episode: string[]
    url: string
    created: string
  }
]

export type AllCharacters = {
  info: {
    count: number
    pages: number
    next: string
    prev: string | null
  }
  results: [
    {
      id: number
      name: string
      status: string
      species: string
      type: string
      gender: string
      origin: {
        name: string
        url: string
      }
      location: {
        name: string
        url: string
      }
      image: string
      episode: string[]
      url: string
      created: string
    }
  ]
}

export type Location = {
  id: number
  name: string
  type: string
  dimension: string
  residents: string[]
  url: string
  created: string
}

export type AllLocations = {
  info: {
    count: number
    pages: number
    next: string
    prev: string | null
  }
  results: [
    {
      id: number
      name: string
      type: string
      dimension: string
      residents: string[]
      url: string
      created: string
    }
  ]
}